Fix for "erresc: unknown str ESC]8" stderr spam
This is a single line fix for the endless standard stderr IO spam
generated by the fact that st has not implemented the xterm hyperlink
extension.

Some context:
- st does not properly support the hyperlink extension/proposal that some other
xterm based terminal emulators support.
- We currently print an error whenever the specific escape
codes are encountered.
- Due to this, whenever I open a terminal my log files are endlessly
flooded with the message: `erresc: unknown str ESC]8`.

The fix seems to work on multiple different builds of st (I've tested
two so far). I've tested the `siduck/snazzy-terminal` build, and the
`mdrdotx/st` build.

I've been using this patch for two weeks or so and I haven't noticed
any problems so I think it is okay.